split import of the two stores to prevent circular dependency
