+++ src/app/store/observation/index.ts
@@ -1,0 +1,3 @@
+export * as observationEffects from './observation.effects';
+export { fromObservation } from './observation.store';
+export { ObservationInfo } from './observation.actions';
