+++ src/app/store/router/index.ts
@@ -1,0 +1,2 @@
+export { fromRouter } from './router.store';
+export { RouterSerializer } from './router-serializer';
